+++ packages/core/src/components/DropdownNew/Dropdown.types.ts
@@ -27,6 +27,10 @@ interface MultiSelectSpecifics<Item extends BaseListItemData<Record<string, unkn
    * The default selected values for multi-select.
    */
   defaultValue?: Item[];
+  /**
+   * The controlled selected values for multi-select.
+   */
+  value?: Item[];
 }
 
 interface SingleSelectSpecifics<Item extends BaseListItemData<Record<string, unknown>>> {
@@ -50,6 +54,10 @@ interface SingleSelectSpecifics<Item extends BaseListItemData<Record<string, unk
    * The default selected value for single-select.
    */
   defaultValue?: Item;
+  /**
+   * The controlled selected value for single-select.
+   */
+  value?: Item;
 }
 
 export type BaseDropdownProps<Item extends BaseListItemData<Record<string, unknown>>> = VibeComponentProps & {
@@ -85,6 +93,15 @@ export type BaseDropdownProps<Item extends BaseListItemData<Record<string, unkno
    * The function to call to render an option.
    */
   optionRenderer?: (option: BaseListItemData<Item>) => React.ReactNode;
+  /**
+   * The function to call to render the menu.
+   */
+  menuRenderer?: (props: {
+    children: React.ReactNode;
+    filteredOptions: ListGroup<Item>[];
+    selectedItems: Item[];
+    getItemProps: (options: any) => Record<string, unknown>;
+  }) => React.ReactNode;
   /**
    * The message to display when there are no options.
    */