+++ packages/core/src/components/DropdownNew/modes/DropdownComboboxController.tsx
@@ -14,6 +14,7 @@ const DropdownComboboxController = <Item extends BaseListItemData<Record<string,
     autoFocus,
     closeMenuOnSelect = true,
     defaultValue,
+    value,
     inputValue: inputValueProp,
     onChange,
     onInputChange,
@@ -30,7 +31,10 @@ const DropdownComboboxController = <Item extends BaseListItemData<Record<string,
     onBlur,
     onKeyDown,
     onClear,
-    onOptionRemove
+    onOptionRemove,
+    size = "medium",
+    readOnly,
+    disabled
   } = props;
 
   const [isFocused, setIsFocused] = useState(false);
@@ -53,6 +57,7 @@ const DropdownComboboxController = <Item extends BaseListItemData<Record<string,
     autoFocus,
     closeMenuOnSelect,
     defaultValue as Item,
+    value as Item,
     inputValueProp,
     onChange,
     onInputChange,
@@ -78,6 +83,7 @@ const DropdownComboboxController = <Item extends BaseListItemData<Record<string,
     getInputProps: (inputOptions?: any) => {
       return hookGetInputProps!({
         ...(inputOptions || {}),
+        disabled: readOnly || disabled,
         onFocus: (event: React.FocusEvent<HTMLInputElement>) => {
           setIsFocused(true);
           onFocus?.(event as any);
@@ -108,7 +114,8 @@ const DropdownComboboxController = <Item extends BaseListItemData<Record<string,
     clearable,
     searchable,
     multi,
-    closeMenuOnSelect
+    closeMenuOnSelect,
+    size
   };
   return <DropdownWrapperUI contextValue={contextValue} dropdownRef={dropdownRef} />;
 };