+++ packages/mcp/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.3](https://github.com/mondaycom/vibe/compare/@vibe/mcp@0.1.2...@vibe/mcp@0.1.3) (2025-05-25)
+
+
+### Bug Fixes
+
+* **mcp:** integrate metadata fetching from unpkg (with cache) instead of consuming @vibe/core ([#2905](https://github.com/mondaycom/vibe/issues/2905)) ([ef1dcd5](https://github.com/mondaycom/vibe/commit/ef1dcd59a8d0f3fff327658a8323ee92a2099312))
+
+
+
+
+
+## [0.1.2](https://github.com/mondaycom/vibe/compare/@vibe/mcp@0.1.1...@vibe/mcp@0.1.2) (2025-05-21)
+
+
+### Bug Fixes
+
+* **mcp:** handle metadata inside mcp instead of consuming @vibe/core as dep ([#2898](https://github.com/mondaycom/vibe/issues/2898)) ([9920a86](https://github.com/mondaycom/vibe/commit/9920a869204a077ef9924ed6c011f9b48417d8ce))
+
+
+
+
+
 ## [0.1.1](https://github.com/mondaycom/vibe/compare/@vibe/mcp@0.1.0...@vibe/mcp@0.1.1) (2025-05-21)